@@ -1,4 +1,4 @@
-package dal
+package gql
 
 import (
 	"fmt"
@@ -11,7 +11,7 @@ import (
 	"github.com/graphql-go/graphql/language/ast"
 	"github.com/mitchellh/mapstructure"
 
-	"github.com/supasheet/dal/internal/dbt"
+	"github.com/supasheet/dal/internal/dal"
 	"github.com/supasheet/dal/internal/warehouse"
 )
 
@@ -42,45 +42,7 @@ var (
 	})
 )
 
-// Build a GraphQL schema from a list of of dbt nodes.
-func BuildSchema(w warehouse.Client, nodes []dbt.Node) *graphql.Schema {
-	fields := buildRoot(w, nodes)
-	rootQuery := graphql.ObjectConfig{Name: "RootQuery", Fields: fields}
-	schemaConfig := graphql.SchemaConfig{Query: graphql.NewObject(rootQuery)}
-
-	schema, err := graphql.NewSchema(schemaConfig)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	return &schema
-}
-
-func buildRoot(w warehouse.Client, nodes []dbt.Node) graphql.Fields {
-	fields := make(graphql.Fields)
-	for _, node := range nodes {
-		fields[node.Name] = &graphql.Field{
-			Description: node.Description,
-			Type:        graphql.NewList(buildType(node)),
-			Resolve:     buildResolver(w, node),
-			Args: graphql.FieldConfigArgument{
-				"limit": &graphql.ArgumentConfig{
-					Type:        graphql.Int,
-					Description: "Limit",
-				},
-				"offset": &graphql.ArgumentConfig{
-					Type:        graphql.Int,
-					Description: "Offset",
-				},
-				"filter": buildFilter(node),
-				"sort":   buildSort(node),
-			},
-		}
-	}
-	return fields
-}
-
-func buildFilter(node dbt.Node) *graphql.ArgumentConfig {
+func buildFilter(model *dal.Model) *graphql.ArgumentConfig {
 	opFields := graphql.InputObjectConfigFieldMap{}
 	for _, op := range []string{"eq", "neq", "lt", "gt", "lte", "gte"} {
 		opFields[op] = &graphql.InputObjectFieldConfig{
@@ -89,11 +51,11 @@ func buildFilter(node dbt.Node) *graphql.ArgumentConfig {
 		}
 	}
 	iocfm := graphql.InputObjectConfigFieldMap{}
-	for _, col := range node.Columns {
+	for _, col := range model.Columns {
 		iocfm[col.Name] = &graphql.InputObjectFieldConfig{
 			Type: graphql.NewInputObject(
 				graphql.InputObjectConfig{
-					Name:   fmt.Sprintf("filter_%s_%s", node.Name, col.Name),
+					Name:   fmt.Sprintf("filter_%s_%s", model.Name, col.Name),
 					Fields: opFields,
 				},
 			),
@@ -102,16 +64,16 @@ func buildFilter(node dbt.Node) *graphql.ArgumentConfig {
 
 	return &graphql.ArgumentConfig{
 		Type: graphql.NewInputObject(graphql.InputObjectConfig{
-			Name:   fmt.Sprintf("%s_filter", node.Name),
+			Name:   fmt.Sprintf("%s_filter", model.Name),
 			Fields: iocfm,
 		}),
 		Description: "Filter",
 	}
 }
 
-func buildSort(node dbt.Node) *graphql.ArgumentConfig {
+func buildSort(model *dal.Model) *graphql.ArgumentConfig {
 	iocfm := graphql.InputObjectConfigFieldMap{}
-	for _, col := range node.Columns {
+	for _, col := range model.Columns {
 		iocfm[col.Name] = &graphql.InputObjectFieldConfig{
 			Type: dirEnum,
 		}
@@ -119,41 +81,15 @@ func buildSort(node dbt.Node) *graphql.ArgumentConfig {
 
 	return &graphql.ArgumentConfig{
 		Type: graphql.NewInputObject(graphql.InputObjectConfig{
-			Name:   fmt.Sprintf("%s_sort", node.Name),
+			Name:   fmt.Sprintf("%s_sort", model.Name),
 			Fields: iocfm,
 		}),
 		Description: "Sort",
 	}
 }
 
-func buildType(node dbt.Node) *graphql.Object {
-	fields := make(graphql.Fields)
-	for _, col := range node.Columns {
-		fields[col.Name] = &graphql.Field{
-			Type:        graphql.String,
-			Description: col.Description,
-		}
-	}
-
-	return graphql.NewObject(graphql.ObjectConfig{
-		Name:   node.Name,
-		Fields: fields,
-	})
-}
-
-//type Filter struct {
-//	Op  string      `json:"op"`
-//	Val interface{} `json:"val"`
-//}
-//
-//func (f Filter) Fragment() goqu.Ex {
-//	return goqu.Ex{
-//		f.Field: goqu.Op{f.Op: f.Val},
-//	}
-//}
-
-func parseFilter(f interface{}) (map[string]map[string]interface{}, error) {
-	var filter map[string]map[string]interface{}
+func parseFilter(f any) (map[string]map[string]any, error) {
+	var filter map[string]map[string]any
 	config := &mapstructure.DecoderConfig{
 		Metadata: nil,
 		Result:   &filter,
@@ -173,10 +109,10 @@ func parseFilter(f interface{}) (map[string]map[string]interface{}, error) {
 	return filter, nil
 }
 
-func buildResolver(w warehouse.Client, node dbt.Node) graphql.FieldResolveFn {
-	return func(p graphql.ResolveParams) (interface{}, error) {
+func buildResolver(w warehouse.Client, model *dal.Model) graphql.FieldResolveFn {
+	return func(p graphql.ResolveParams) (any, error) {
 		// Generate the SQL query
-		q := dialect.From(node.Name).Select(getSelectedFields(p)...)
+		q := dialect.From(model.Name).Select(getSelectedFields(model.PrimaryKey, p)...)
 
 		// Handle filter
 		if f, ok := p.Args["filter"]; ok {
@@ -197,7 +133,7 @@ func buildResolver(w warehouse.Client, node dbt.Node) graphql.FieldResolveFn {
 		// Handle sort
 		if o, ok := p.Args["sort"]; ok {
 			var oes []exp.OrderedExpression
-			for field, dir := range o.(map[string]interface{}) {
+			for field, dir := range o.(map[string]any) {
 				c := goqu.C(field)
 				var oe exp.OrderedExpression
 				if dir == "asc" {
@@ -247,18 +183,10 @@ func cleanQuery(query string) string {
 	return string(cleaned)
 }
 
-func cleanQueryString(query string) string {
-	var cleaned []rune
-	for _, r := range query {
-		if r != 0 {
-			cleaned = append(cleaned, r)
-		}
-	}
-	return string(cleaned)
-}
-
-// Returns the list of request fields listed under provider selection path in the Graphql query.
-func getSelectedFields(p graphql.ResolveParams) []interface{} {
+// Returns the list of requested fields from the current part of the query.  It
+// requires us to specify the pk to ensure that it's always present in the
+// result.
+func getSelectedFields(pk string, p graphql.ResolveParams) []any {
 	var selectionPath []string
 	for _, part := range p.Info.Path.AsArray() {
 		selectionPath = append(selectionPath, part.(string))
@@ -275,7 +203,13 @@ func getSelectedFields(p graphql.ResolveParams) []interface{} {
 				fields = make([]*ast.Field, 0)
 
 				for _, selection := range selections {
-					fields = append(fields, selection.(*ast.Field))
+					// We only want 'raw' fields to be included in this sql
+					// query. Foreign Keys are resolved differently, so we only
+					// add it in if it does not have a selection set of its
+					// own.
+					if selection.GetSelectionSet() == nil {
+						fields = append(fields, selection.(*ast.Field))
+					}
 				}
 
 				found = true
@@ -285,14 +219,16 @@ func getSelectedFields(p graphql.ResolveParams) []interface{} {
 		}
 
 		if !found {
-			return []interface{}{}
+			return []any{}
 		}
 	}
 
-	var collect []interface{}
-
+	collect := []any{pk}
 	for _, field := range fields {
-		collect = append(collect, field.Name.Value)
+		n := field.Name.Value
+		if n != pk {
+			collect = append(collect, n)
+		}
 	}
 
 	return collect
