@@ -3,11 +3,12 @@ package dbt
 import (
 	"log"
 	"os"
+	"path/filepath"
 
 	"gopkg.in/yaml.v3"
 )
 
-type ProjectConfig struct {
+type Project struct {
 	Name              string                 `json:"name"`
 	ConfigVersion     int                    `json:"config-version"`
 	Version           string                 `json:"version"`
@@ -30,18 +31,43 @@ type ModelConfig struct {
 	} `json:"staging"`
 }
 
-func Project() ProjectConfig {
+func LoadProject() *Project {
 	f, err := os.Open("./dbt_project.yml")
 	if err != nil {
 		log.Fatal(err)
 	}
 	defer f.Close()
 
-	var pc ProjectConfig
+	var pc Project
 	err = yaml.NewDecoder(f).Decode(&pc)
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	return pc
+	return &pc
+}
+
+func (p *Project) LoadProfile() Profile {
+	// Find the home directory
+	home, err := os.UserHomeDir()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	// Open the profile file
+	f, err := os.Open(filepath.Join(home, ".dbt", "profiles.yml"))
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer f.Close()
+
+	// Load the profiles
+	var profiles Profiles
+	err = yaml.NewDecoder(f).Decode(&profiles)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	// Return the selected profile
+	return profiles[p.Profile]
 }
