@@ -2,21 +2,49 @@ package warehouse
 
 import (
 	"database/sql"
+	"regexp"
 
 	"github.com/snowflakedb/gosnowflake"
+	"github.com/supasheet/dal/internal/dal"
 )
 
+type SnowflakeCredentials struct {
+	AccountId string `json:"account_id"`
+	User      string `json:"user"`
+	Password  string `json:"password"`
+	Database  string `json:"database"`
+	Schema    string `json:"schema"`
+	Warehouse string `json:"warehouse"`
+}
+
+func (sfc SnowflakeCredentials) ConnString() (string, error) {
+	cfg := gosnowflake.Config{
+		Account:   sfc.AccountId,
+		User:      sfc.User,
+		Password:  sfc.Password,
+		Database:  sfc.Database,
+		Schema:    sfc.Schema,
+		Warehouse: sfc.Warehouse,
+	}
+	dsn, err := gosnowflake.DSN(&cfg)
+	if err != nil {
+		return "", err
+	}
+	return dsn, nil
+}
+
 type SnowflakeClient struct {
-	db *sql.DB
+	db    *sql.DB
+	creds SnowflakeCredentials
 }
 
-func NewSnowflake() *SnowflakeClient {
-	return &SnowflakeClient{}
+func NewSnowflake(sfc SnowflakeCredentials) *SnowflakeClient {
+	return &SnowflakeClient{creds: sfc}
 }
 
-func (sc *SnowflakeClient) Connect(sfc SnowflakeCredentials) error {
+func (sc *SnowflakeClient) Connect() error {
 	// Create the db instance
-	dsn, err := sfc.ConnString()
+	dsn, err := sc.creds.ConnString()
 	if err != nil {
 		return err
 	}
@@ -65,27 +93,15 @@ func (sc *SnowflakeClient) Run(query string) (Records, error) {
 	return records, nil
 }
 
-type SnowflakeCredentials struct {
-	AccountId string `json:"account_id"`
-	User      string `json:"user"`
-	Password  string `json:"password"`
-	Database  string `json:"database"`
-	Schema    string `json:"schema"`
-	Warehouse string `json:"warehouse"`
+func (sc *SnowflakeClient) MapType(t string) dal.Scalar {
+	return snowflakeDataTypeMatcher.match(t)
 }
 
-func (sfc SnowflakeCredentials) ConnString() (string, error) {
-	cfg := gosnowflake.Config{
-		Account:   sfc.AccountId,
-		User:      sfc.User,
-		Password:  sfc.Password,
-		Database:  sfc.Database,
-		Schema:    sfc.Schema,
-		Warehouse: sfc.Warehouse,
-	}
-	dsn, err := gosnowflake.DSN(&cfg)
-	if err != nil {
-		return "", err
-	}
-	return dsn, nil
+var snowflakeDataTypeMatcher = &dataTypeMatcher{
+	id:       regexp.MustCompile("a^"),
+	int:      regexp.MustCompile("(?i)(FIXED|INT|NUMBER|NUMERIC|DECIMAL)"),
+	float:    regexp.MustCompile("(?i)(REAL|FLOAT|DOUBLE)"),
+	boolean:  regexp.MustCompile("(?i)(BOOLEAN)"),
+	string:   regexp.MustCompile("(?i)(CHAR|STRING|TEXT)"),
+	dateTime: regexp.MustCompile("(?i)(TIME|DATE)"),
 }
